+++ amplifier.cpp
@@ -12,11 +12,11 @@ using std::endl;
 using std::ofstream;
 using std::setprecision;
 
-double amplifier::amp(double Vin, double dt, bool ind)
+long double amplifier::amp(long double Vin, long double dt, bool ind)
 {
 
 //linking op-amp cables
-double vtemp = 0.0;
+long double vtemp = 0.0;
 
 vtemp = AA.ampID(Vin,dt);
 vtemp = BB.amp(vtemp,dt);
@@ -31,7 +31,7 @@ return vtemp;
 }
 
 
-double amplifier::ampID(double Vin, double dt, bool ind, bool WO, ofstream  & out)
+long double amplifier::ampID(long double Vin, long double dt, bool ind, bool WO, ofstream  & out)
 {
 
 //"amp_data.txt" contains the signal values after each amplification
@@ -39,7 +39,7 @@ double amplifier::ampID(double Vin, double dt, bool ind, bool WO, ofstream  & ou
 //(W)rite(O)utput WO.	
 
 //linking op-amp cables
-double vtemp = 0.0;
+long double vtemp = 0.0;
 if(WO){out<< Vin << "\t";}
 vtemp = AA.ampID(Vin,dt);
 if(WO){out<< vtemp << "\t";}
@@ -61,9 +61,9 @@ if(WO){out<< vtemp << endl;}
 return vtemp;
 }
 
-double amplifier::LP(double Vin, double dt){
+long double amplifier::LP(long double Vin, long double dt){
 
-	double vtemp = 0.0;
+	long double vtemp = 0.0;
 	vtemp = AA.ampID(Vin,dt);
 	return vtemp;
